update and support latest AutomaticWaterMasking
1) rename AreaKMLFromOSMDataCreator to OSMXMLParser
2) AutomaticWaterMasking now utilizes doubles instead of decimal
(significant speed increase for certain tiles)

+++ FSEarthTiles/FSEarthMasksInternalDLL/MasksTexture.cs
@@ -4748,9 +4748,9 @@ namespace FSEarthMasksInternalDLL
 
                 foreach (MaskingPolys polys in allMaskingPolys)
                 {
-                    decimal pixelsPerLon = (decimal)(Convert.ToDouble(MasksConfig.mAreaPixelCountInX) / (MasksConfig.mAreaSECornerLongitude - MasksConfig.mAreaNWCornerLongitude));
-                    decimal pixelsPerLat = (decimal)(Convert.ToDouble(MasksConfig.mAreaPixelCountInY) / (MasksConfig.mAreaNWCornerLatitude - MasksConfig.mAreaSECornerLatitude));
-                    AutomaticWaterMasking.Point NW = new AutomaticWaterMasking.Point((decimal)NWLon, (decimal)NWLat);
+                    double pixelsPerLon = Convert.ToDouble(MasksConfig.mAreaPixelCountInX) / (MasksConfig.mAreaSECornerLongitude - MasksConfig.mAreaNWCornerLongitude);
+                    double pixelsPerLat = Convert.ToDouble(MasksConfig.mAreaPixelCountInY) / (MasksConfig.mAreaNWCornerLatitude - MasksConfig.mAreaSECornerLatitude);
+                    AutomaticWaterMasking.Point NW = new AutomaticWaterMasking.Point(NWLon, NWLat);
                     b = new SolidBrush(Color.Black);
                     CommonFunctions.DrawPolygons(bmp, g, b, pixelsPerLon, pixelsPerLat, NW, polys.coastWaterPolygons);
                     for (int i = 0; i < polys.inlandPolygons.Length; i++)
